Minerva: Fusion-related functions

Minerva: Fusion-related functions

Index

PlasmaFunctions

Functions

Index

Minerva.Lib.PlasmaFunctions.get_DT_reactionRate

Minerva.Lib.PlasmaFunctions.get_TProfile

Minerva.Lib.PlasmaFunctions.get_nAveraged

Minerva.Lib.PlasmaFunctions.get_nAveraged2

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk_2018

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius_2018

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor_2018

Minerva.Lib.PlasmaFunctions.get_nAvg_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor_2018

Minerva.Lib.PlasmaFunctions.get_nProfile

Minerva.Lib.PlasmaFunctions.get_reactionRates

PlasmaFunctions

Functions

Minerva.Lib.PlasmaFunctions.get_nProfile — Function

get_nProfile(n0, np, a, ap, αn, βn)

returns the plasma density profile.

Arguments

n0:

np:

a:

ap:

αn:

βn:

Minerva.Lib.PlasmaFunctions.get_TProfile — Function

get_TProfile(n0, np, a, ap, αn, βn)

returns the plasma temperature profile.

Arguments

T0:

Tp:

a:

ap:

αt:

βt:

Minerva.Lib.PlasmaFunctions.get_nAveraged — Function

get_nAveraged(n0, np, a, ap, αn, βn)

Returns the line-average plasma density (average over the radius). Takes as parameters the coefficients of a profile with the shape [EQUATION]

Arguments

T0:

Tp:

a:

ap:

αt:

βt:

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius — Function

get_nAveraged_radius(n0, np, a, ap, αn, βn)

Returns the line-average plasma density (average over the radius). Takes as parameters the coefficients of a profile with the shape [EQUATION]

Arguments

T0:

Tp:

a:

ap:

αt:

βt:

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk — Function

get_nAveraged_disk(n0, np, a, ap, αn, βn)

Returns the area-average plasma density (average over the cross-section). Takes as parameters the coefficients of a profile with the shape [EQUATION]

Arguments

T0:

Tp:

a:

ap:

αt:

βt:

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius_2018 — Function

get_nAveraged_radius_2018(n)

Returns the line-average plasma density (average over the radius).

Arguments

n: function representing the density profile. n(x), where 0 <= x <= 1

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk_2018 — Function

get_nAveraged_disk_2018(n)

Returns the area-average plasma density (average over the cross-section).

Arguments

n: function representing the density profile. n(x), where 0 <= x <= 1

Minerva.Lib.PlasmaFunctions.get_nAveraged2 — Function

get_nAveraged2(n0, np, a, ap, αn, βn, xaz)

Returns the line-average plasma density (average over the radius) over the Active Zone.

Arguments

'n0':

'np':

'a':

'ap':

'αn':

'βn':

'xaz':

Minerva.Lib.PlasmaFunctions.get_nAvg_powerCorrectionFactor — Function

get_nAvg_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor — Function

function get_nAvg_radius_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor — Function

get_nAvg_disk_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor_2018 — Function

get_nAvg_radius_powerCorrectionFactor_2018(n, xaz)

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor_2018 — Function

get_nAvg_disk_powerCorrectionFactor_2018(n, T, xaz)

Minerva.Lib.PlasmaFunctions.get_reactionRates — Function

get_reactionRates(xaz::Float64, k::Float64, a::Float64, R::Float64, T::Function)

Minerva.Lib.PlasmaFunctions.get_DT_reactionRate — Function

get_DT_reactionRate(xaz::Float64, k::Float64, a::Float64, R::Float64, T::Function, n)