Minerva: Components

Minerva: Components

Index

Atomic

Atomic Blocks

Index

Minerva.Lib.Base.Atomic.AFPInlet

Minerva.Lib.Base.Atomic.AFPInlet_HDTO

Minerva.Lib.Base.Atomic.AFPInlet_O

Minerva.Lib.Base.Atomic.AFPInlet_OC

Minerva.Lib.Base.Atomic.AFPInlet_OCN

Minerva.Lib.Base.Atomic.AFPOutlet

Minerva.Lib.Base.Atomic.AFPOutlet_HDTO

Minerva.Lib.Base.Atomic.AFPOutlet_O

Minerva.Lib.Base.Atomic.AFPOutlet_OC

Minerva.Lib.Base.Atomic.AFPOutlet_OCN

Minerva.Lib.Base.Atomic.FluidOpenBC

Minerva.Lib.Base.Atomic.FluidOpenBC_HDTO

Minerva.Lib.Base.Atomic.FluidOpenBC_OCN

Minerva.Lib.Base.Atomic.FluidOpenBC_withO

Minerva.Lib.Base.Atomic.HDTO_to_outlet

Minerva.Lib.Base.Atomic.MFP

Minerva.Lib.Base.Atomic.MixerThreeWay

Minerva.Lib.Base.Atomic.MixerThreeWayOCN

Minerva.Lib.Base.Atomic.MixerTwoWay

Minerva.Lib.Base.Atomic.MixerTwoWayOCN

Minerva.Lib.Base.Atomic.Plasma

Minerva.Lib.Base.Atomic.TorusMixer

Minerva.Lib.Base.Atomic.TorusMixer_withFuelEff

Minerva.Lib.Base.Atomic.inlet2_to_outletOCN

Minerva.Lib.Base.Atomic.inlet3_to_outletOCN

Minerva.Lib.Base.Atomic.inlet_to_outlet2

Atomic

Atomic Blocks

Minerva.Lib.Base.Atomic.AFPInlet — Constant

AFPInlet(;name)

Inlet connector for species H, D, T, He, Xe and one inert 'I'.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xI(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPOutlet — Constant

AFPOutlet(;name)

Outlet connector for species H, D, T, He, Xe and one inert 'I'.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xI(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPInlet_O — Constant

AFPInlet_O(;name)

Inlet connector for species H, D, T, He, Xe, O and one inert 'I'.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xI(t): [-] atomic fraction

xO(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPOutlet_O — Constant

AFPOutlet_O(;name)

Outlet connector for species H, D, T, He, Xe, O and one inert 'I'.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xI(t): [-] atomic fraction

xO(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPInlet_HDTO — Constant

AFPInlet_HDTO(;name)

Inlet connector for species H, D, T, O.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xO(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPOutlet_HDTO — Constant

AFPOutlet_HDTO(;name)

Outlet connector for species H, D, T, O.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xO(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPInlet_OC — Constant

AFPInlet_OC(;name)

Inlet connector for species H, D, T, O, He, Xe, C, 'I'

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xO(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xC(t): [-] atomic fraction

xI(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPOutlet_OC — Constant

AFPOutlet_OC(;name)

Outlet connector for species H, D, T, O, He, Xe, C, 'I'

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xO(t): [-] atomic fraction

xHe(t): [-] atomic fraction

xXe(t): [-] atomic fraction

xC(t): [-] atomic fraction

xI(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPInlet_OCN — Constant

AFPInlet_OCN(;name)

Inlet connector for species H, D, T, 'I', O, C, N.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xI(t): [-] atomic fraction

xO(t): [-] atomic fraction

xC(t): [-] atomic fraction

xN(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.AFPOutlet_OCN — Constant

AFPOutlet_OCN(;name)

Outlet connector for species H, D, T, 'I', O, C, N.

States:

G(t): [$\mathrm{s^{-1}}$] flow rate

xH(t): [-] atomic fraction

xD(t): [-] atomic fraction

xT(t): [-] atomic fraction

xI(t): [-] atomic fraction

xO(t): [-] atomic fraction

xC(t): [-] atomic fraction

xN(t): [-] atomic fraction

p(t): [$\mathrm{Pa}$] pressure

T(t): [$\mathrm{K}$] temperature

Minerva.Lib.Base.Atomic.inlet_to_outlet2 — Constant

inlet_to_outlet2(;name)

Component to pass from AFPOutlet to AFPInlet_O. Species 'O' is added at the outlet.

Components:

a: AFPInlet [inlet]

b: AFPOutlet_O [outlet]

Minerva.Lib.Base.Atomic.HDTO_to_outlet — Constant

HDTO_to_outlet(;name)

Component to pass from AFPOutlet_HDTO to AFPInlet. Species 'O' is removed at the outlet. Species 'He, Xe, I' are added at the outlet.

Components:

a: AFPInlet_HDTO [inlet]

b: AFPOutlet [outlet]

Minerva.Lib.Base.Atomic.inlet3_to_outletOCN — Constant

inlet3_to_outletOCN(;name)

Component to pass from AFPOutletOC to AFPInletOCN. Species 'N' is added at the outlet.

Components:

a: AFPInlet_OC [inlet]

b: AFPOutlet_OCN [outlet]

Minerva.Lib.Base.Atomic.inlet2_to_outletOCN — Constant

inlet2_to_outletOCN(;name)

Component to pass from AFPOutletO to AFPInletOCN. Species 'C, N' are added at the outlet.

Components:

a: AFPInlet_OC [inlet]

b: AFPOutlet_OCN [outlet]

Minerva.Lib.Base.Atomic.Plasma — Constant

Plasma(;name)

Torus model with with an inlet port a, one outlet port b, and a fluid flow with 4 species.

States:

xH(t): [-] outlet H molar fraction

xD(t): [-] outlet D molar fraction

xT(t): [-] otulet T molar fraction

xHe(t): [-] outlet He molar fraction

xI(t): [-] outlet He molar fraction

G_H(t): [$\mathrm{s^{-1}}$] protium production rate

G_alpha(t): [$\mathrm{s^{-1}}$] $\alpha$ particle production rate

DNH(t): [$\mathrm{s^{-1}}$] time derivative of protium population

Connectors:

a: inlet fluid port

b : outlet fluid port

Parameters:

a: inlet fluid port

b : outlet fluid port

τp: [-] particle confinement time

τp_α: [-] alpha particle confinement time

Vₚₗₐₛₘₐ: [-] plasma volume

n: [-] initial fuel density

rR_DD_1: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_DD_2: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_DT: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_DHe3: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_TT: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_THe3_1: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_THe3_2: [$\mathrm{m^6 s^{-1}}$] reaction rate

rR_THe3_3: [$\mathrm{m^6 s^{-1}}$] reaction rate

Wcorrection: [-] density correction factor

Equations:

\[\begin{aligned} V_{pl}\frac{\partial n_H}{\partial t} =\ & x_H^{in} \Gamma^{in} + x_H^{out} \Gamma^{out} + w_c n_D n_{He} R_{DHe^3} + w_c n_D n_D R_{DD-1} + \\ & + w_c n_T n_{He} (R_{THe^3-1} + R_{THe^3-3})\\ V_{pl}\frac{\partial n_D}{\partial t} =\ & x_D^{in} \Gamma^{in} + x_D^{out} \Gamma^{out} + w_c n_T n_{He} R_{THe^3-2} - w_c n_D n_{He} R_{DHe^3} - \\ & - w_c n_D n_T R_{DT} - w_c 2 n_D n_D (R_{DD-1} + R_{DD-2})\\ V_{pl}\frac{\partial n_T}{\partial t} =\ & x_T^{in} \Gamma^{in} + x_T^{out} \Gamma^{out} + w_c n_D n_D R_{DD-1} - w_c n_D n_T R_{DT} - \\ & - w_c 2 n_T n_T R_{TT} - w_c n_T n_{He} (R_{THe^3-1} + R_{THe^3-2} + R_{THe^3-3})\\ V_{pl}\frac{\partial n_{He}}{\partial t} =\ & x_{He}^{in} \Gamma^{in} + x_{He}^{out} \Gamma^{out} + w_c n_D n_D R_{DD-2} + w_c n_D n_T R_{DT} + w_c n_T n_T R_{TT}\\ x_i \Gamma^{out} =\ & - n_i V_{pl} / \tau_p,\ \ \ \ i = H,D,T,He\\ \sum_{i} x_i =\ & 1,\ \ \ \ i = H,D,T,H\\ \end{aligned}\]

Minerva.Lib.Base.Atomic.TorusMixer — Constant

TorusMixer(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Ne

Connectors:

'a_FW': inlet fluid port

'a_plasma': inlet fluid port

'b' : outlet fluid port

Minerva.Lib.Base.Atomic.TorusMixer_withFuelEff — Constant

TorusMixer_withFuelEff(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Ne

Connectors:

'a_FW': inlet fluid port

'a_plasma': inlet fluid port

'b' : outlet fluid port

Minerva.Lib.Base.Atomic.MFP — Constant

MFP(;name)

Metal Foil Pump (MFP) model. It has one inlet port 'a' and two outlet ports 'b1' and 'b2'. It has a fluid flow and 6 species composition. A portion eta of the inlet Q (H,D,T) is sent to the outlet 'b1', the rest + (1-eta)Q is sent to the outlet port 'b2'.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of I

Connectors:

'a': inlet fluid port

'b1' : outlet fluid port

'b2' : outlet fluid port

Minerva.Lib.Base.Atomic.FluidOpenBC — Constant

FluidOpenBC(;name)

Simple BC for a fluid port such that the mass flow 'dm' is non-zero. It is the opposite from an un-connected port.

States:

'dm(t)': [mol s-1] The molar flow rate passing through the 'a' port

'xH(t)': [-] The molar fraction of H2 passing thorugh the 'a' port

'xD(t)': [-] The molar fraction of D2 passing thorugh the 'a' port

'xT(t)': [-] The molar fraction of T2 passing thorugh the 'a' port

'xHe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xXe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xI(t)': [-] The molar fraction of Inert passing thorugh the 'a' port

Connectors:

'a': inlet fluid port

Minerva.Lib.Base.Atomic.FluidOpenBC_withO — Constant

FluidOpenBC_withO(;name)

Simple BC for a fluid port such that the mass flow 'dm' is non-zero. It is the opposite from an un-connected port. THe ports contain also Oxygen 'O'

States:

'dm(t)': [mol s-1] The molar flow rate passing through the 'a' port

'xH(t)': [-] The molar fraction of H2 passing thorugh the 'a' port

'xD(t)': [-] The molar fraction of D2 passing thorugh the 'a' port

'xT(t)': [-] The molar fraction of T2 passing thorugh the 'a' port

'xHe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xXe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xI(t)': [-] The molar fraction of Inert passing thorugh the 'a' port

'xO(t)': [-] The molar fraction of O passing thorugh the 'a' port

Connectors:

'a': inlet fluid port

Minerva.Lib.Base.Atomic.FluidOpenBC_OCN — Constant

FluidOpenBC_OCN(;name)

Simple BC for a fluid port such that the mass flow 'dm' is non-zero. It is the opposite from an un-connected port. THe ports contain also Oxygen 'O', 'C', 'N'

States:

'dm(t)': [mol s-1] The molar flow rate passing through the 'a' port

'xH(t)': [-] The molar fraction of H2 passing thorugh the 'a' port

'xD(t)': [-] The molar fraction of D2 passing thorugh the 'a' port

'xT(t)': [-] The molar fraction of T2 passing thorugh the 'a' port

'xHe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xXe(t)': [-] The molar fraction of He passing thorugh the 'a' port

'xI(t)': [-] The molar fraction of Inert passing thorugh the 'a' port

'xO(t)': [-] The molar fraction of O passing thorugh the 'a' port

'xC(t)': [-] The molar fraction of C passing thorugh the 'a' port

'xN(t)': [-] The molar fraction of N passing thorugh the 'a' port

Connectors:

'a': inlet fluid port

Minerva.Lib.Base.Atomic.FluidOpenBC_HDTO — Constant

FluidOpenBC_HDTO(;name)

Simple BC for a fluid port such that the mass flow 'dm' is non-zero. It is the opposite from an un-connected port.

States:

'dm(t)': [mol s-1] The flow rate passing through the 'a' port

'xH(t)': [-] The fraction of H passing thorugh the 'a' port

'xD(t)': [-] The fraction of D passing thorugh the 'a' port

'xT(t)': [-] The fraction of T passing thorugh the 'a' port

'xO(t)': [-] The fraction of O passing thorugh the 'a' port

Connectors:

'a': inlet fluid port

Minerva.Lib.Base.Atomic.MixerTwoWay — Constant

MixerTwoWay(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Inert

Connectors:

'a1': inlet fluid port

'a2': inlet fluid port

'b' : outlet fluid port

Minerva.Lib.Base.Atomic.MixerTwoWayOCN — Constant

MixerTwoWayOCN(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Inert

'xO(t)': [-] The inlet molar fraction of O

'xC(t)': [-] The inlet molar fraction of C

'xN(t)': [-] The inlet molar fraction of N

Connectors:

'a1': inlet fluid port

'a2': inlet fluid port

'b' : outlet fluid port

Minerva.Lib.Base.Atomic.MixerThreeWay — Constant

MixerThreeWay(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Inert

Connectors:

'a1': inlet fluid port

'a2': inlet fluid port

'a3': inlet fluid port

'b' : outlet fluid port

Minerva.Lib.Base.Atomic.MixerThreeWayOCN — Constant

MixerThreeWayOCN(;name)

Mixer volume with two inlet ports 'a1' and 'a2' and one outlet port 'b', a fluid flow and 6 species composition.

States:

'dm(t)': [mol s-1] The inlet molar flow rate

'xH(t)': [-] The inlet molar fraction of H2

'xD(t)': [-] The inlet molar fraction of D2

'xT(t)': [-] The inlet molar fraction of T2

'xHe(t)': [-] The inlet molar fraction of He

'xXe(t)': [-] The inlet molar fraction of He

'xI(t)': [-] The inlet molar fraction of Inert

'xO(t)': [-] The inlet molar fraction of O

'xC(t)': [-] The inlet molar fraction of C

'xN(t)': [-] The inlet molar fraction of N

Connectors:

'a1': inlet fluid port

'a2': inlet fluid port

'a3': inlet fluid port

'b' : outlet fluid port