Minerva: Fusion-related functions

Minerva: Fusion-related functions

Index

PlasmaFunctions

Functions

Index

Minerva.Lib.PlasmaFunctions.get_DT_reactionRate

Minerva.Lib.PlasmaFunctions.get_TProfile

Minerva.Lib.PlasmaFunctions.get_nAveraged

Minerva.Lib.PlasmaFunctions.get_nAveraged2

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk_2018

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius_2018

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor_2018

Minerva.Lib.PlasmaFunctions.get_nAvg_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor_2018

Minerva.Lib.PlasmaFunctions.get_nProfile

Minerva.Lib.PlasmaFunctions.get_reactionRates

PlasmaFunctions

Functions

Minerva.Lib.PlasmaFunctions.get_nProfile — Function

get_nProfile(n0, np, a, ap, αn, βn)

returns the plasma density profile

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

Minerva.Lib.PlasmaFunctions.get_TProfile — Function

get_TProfile(n0, np, a, ap, αn, βn)

returns the plasma temperature profile.

\[T(x, T_p, T_0) = T_p + (T_0 - T_p) \cdot [1 - (\frac{a}{ap})^{\beta t} \cdot x^{\beta t}]^{\alpha t}\]

Arguments

T0: temperature at the plasma center [keV]

Tp: temperature at the pedestel [keV]

a: coefficient [-]

ap: coefficient [-]

αt: coefficient [-]

βt: coefficient [-]

Minerva.Lib.PlasmaFunctions.get_nAveraged — Function

get_nAveraged(n0, np, a, ap, αn, βn)

Returns the line-average plasma density (average over the radius). Takes as parameters the coefficients of a profile with the shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius — Function

get_nAveraged_radius(n0, np, a, ap, αn, βn)

Returns the line-average plasma density (average over the radius). Takes as parameters the coefficients of a profile with the shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk — Function

get_nAveraged_disk(n0, np, a, ap, αn, βn)

Returns the area-average plasma density (average over the cross-section). Takes as parameters the coefficients of a profile with the shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

Minerva.Lib.PlasmaFunctions.get_nAveraged_radius_2018 — Function

get_nAveraged_radius_2018(n)

Returns the line-average plasma density (average over the radius).

Arguments

n: function representing the density profile. n(x), where 0 <= x <= 1 [m-3]

Minerva.Lib.PlasmaFunctions.get_nAveraged_disk_2018 — Function

get_nAveraged_disk_2018(n)

Returns the area-average plasma density (average over the cross-section).

Arguments

n: function representing the density profile. n(x), where 0 <= x <= 1 [m-3]

Minerva.Lib.PlasmaFunctions.get_nAveraged2 — Function

get_nAveraged2(n0, np, a, ap, αn, βn, xaz)

Returns the line-average plasma density (average over the radius) over the Active Zone.

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

'xaz': active zone normalized radius [-]

Minerva.Lib.PlasmaFunctions.get_nAvg_powerCorrectionFactor — Function

get_nAvg_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

When computing fusion reactions over pre-integrated reactions rates, a correction factor needs to be used to account for the use of averaged densities. This functions computes and returns this correction factor for computing reactions rates over the active zone.

It takes as input a density profile with the following shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

'xaz': active zone normalized radius [-]

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor — Function

get_nAvg_radius_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

When computing fusion reactions over pre-integrated reactions rates, a correction factor needs to be used to account for the use of averaged densities. This functions computes and returns this correction factor for computing reactions rates over the active zone.

It takes as input a density profile with the following shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Use this function if working with line-averaged densities

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

'xaz': active zone normalized radius [-]

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor — Function

get_nAvg_disk_powerCorrectionFactor(n0, np, a, ap, αn, βn, xaz)

When computing fusion reactions over pre-integrated reactions rates, a correction factor needs to be used to account for the use of averaged densities. This functions computes and returns this correction factor for computing reactions rates over the active zone.

It takes as input a density profile with the following shape:

\[n(x, n_p, n_0) = n_p + (n_0 - n_p) \cdot [1 - (\frac{a}{ap})^2 \cdot x^2]^{\alpha n}\]

Use this function if working with area-averaged densities

Arguments

n0: density at the plasma center [m-3]

np: density at the pedestel [m-3]

a: coefficient [-]

ap: coefficient [-]

αn: coefficient [-]

βn: coefficient [-]

'xaz': active zone normalized radius [-]

Minerva.Lib.PlasmaFunctions.get_nAvg_radius_powerCorrectionFactor_2018 — Function

get_nAvg_radius_powerCorrectionFactor_2018(n, xaz)

When computing fusion reactions over pre-integrated reactions rates, a correction factor needs to be used to account for the use of averaged densities. This functions computes and returns this correction factor for computing reactions rates over the active zone.

It takes as input a density profile expressed as a function of the normalized plasma radius. Use this function if working with line-averaged densities

Arguments

n: density profile as a function of coordinate x, with 0 <= x <= 1 [m-3]

'xaz': active zone normalized radius [-]

Minerva.Lib.PlasmaFunctions.get_nAvg_disk_powerCorrectionFactor_2018 — Function

get_nAvg_disk_powerCorrectionFactor_2018(n, T, xaz)

When computing fusion reactions over pre-integrated reactions rates, a correction factor needs to be used to account for the use of averaged densities. This functions computes and returns this correction factor for computing reactions rates over the active zone.

It takes as input a density and a temperature profiles expressed as a function of the normalized plasma radius. Use this function if working with area-averaged densities

Arguments

n: density profile as a function of coordinate x, with 0 <= x <= 1

T: temperature profile as a function of coordinate x, with 0 <= x <= 1

'xaz': active zone normalized radius

Minerva.Lib.PlasmaFunctions.get_reactionRates — Function

get_reactionRates(xaz::Float64, k::Float64, a::Float64, R::Float64, T::Function)

Integral of reaction rates over the active zone Returns 8 values related to 8 reaction branches:

rR_DD_1: [$\mathrm{m^6 s^{-1}}$] D + D –(50%)–> T + p

rR_DD_2: [$\mathrm{m^6 s^{-1}}$] D + D –(50%)–> He3 + n

rR_DT: [$\mathrm{m^6 s^{-1}}$] D + T –> He4 + n

rR_DHe3: [$\mathrm{m^6 s^{-1}}$] D + He3 –> He4 + p

rR_TT: [$\mathrm{m^6 s^{-1}}$] T + T –> He4 + 2n + 11.3 MeV

rR_THe3_1: [$\mathrm{m^6 s^{-1}}$] T + He3 –(51%)–> He4 + p + n + 12.1 MeV

rR_THe3_2: [$\mathrm{m^6 s^{-1}}$] T + He3 –(43%)–> He4 + D

rR_THe3_3: [$\mathrm{m^6 s^{-1}}$] T + He3 –( 6%)–> He5 + p

Arguments

'xaz': active zone normalized radius [-]

'k': plasma shape factor [-]

'a': minor radius [m]

'R': major radius [m]

'T': temperature profile (function of normalized radius) [keV]

Minerva.Lib.PlasmaFunctions.get_DT_reactionRate — Function

get_DT_reactionRate(xaz::Float64, k::Float64, a::Float64, R::Float64, T::Function, n)

Return the D-T reaction rate [$\mathrm{#reactions s^{-1}}$]

Arguments

'xaz': active zone normalized radius [-]

'k': plasma shape factor [-]

'a': minor radius [m]

'R': major radius [m]

'T': temperature profile (function of normalized radius) [keV]

'n': density profile (function of normalized radius) [m-3]